Optimize out the large string concats
A pprof CPU profile told me that the program was spending most of its
time in... string concatenations?!? Turns out that the calls to memmove
during string concats were quite expensive. So we optimize that out by
creating an array of small strings representing the things we want in
our final string, and building the string in the very end.